% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{parallel_compnetaccess}
\alias{parallel_compnetaccess}
\title{A function to compute distance/cost to point of interest using cost-matrix system in parallel}
\usage{
parallel_compnetaccess(
  cpus,
  lines_obj,
  origins_dt,
  dest_dt,
  cost_matrix = FALSE,
  weight = "time",
  directed = FALSE,
  length_as_weight = TRUE,
  blend_obj = NULL,
  blend_dsn = NULL
)
}
\arguments{
\item{cpus}{an \code{integer}, the number of CPUs to divide process into for parallelization}

\item{lines_obj}{an \code{sfnetwork} object. apply as_\code{sfnetwork()} function to \code{sf}, \code{data.frame} object}

\item{origins_dt}{an \code{sf} object showing the set of origin points}

\item{dest_dt}{on \code{sf} object showing the destimations or points of interest to optimize on}

\item{cost_matrix}{where or not the cost matrix should be returned}

\item{weight}{the variable within the \code{lines_obj} to be used as measure cost measure (e.g time)}

\item{directed}{FALSE if each edge is assumed to be two-way directed.}

\item{length_as_weight}{TRUE if \code{weight} is to be length of edges within the network}

\item{blend_obj}{a \code{character}, for the filepath to .RDS file to a blended network object of network, origin and destination
objects. See \code{sfnetwork::st_network_blend()} for more details of \code{blend_obj}.}

\item{blend_dsn}{a \code{character}, the folder location to store \code{blend_obj} when the function creates it. Otherwise,
leave as NULL.}

\item{parallel_model}{see \code{parallelMap::parallelLapply()} for more details}
}
\description{
A function to compute distance/cost to point of interest using cost-matrix system in parallel
}
