% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{parallel_zonalext}
\alias{parallel_zonalext}
\title{A function to parallel compute zonal stats}
\usage{
parallel_zonalext(
  shp_dt,
  raster_obj,
  summary_fun,
  cpus,
  parallel_mode = "multicore"
)
}
\arguments{
\item{shp_dt}{sf, data.frame polygon/multipolygon object}

\item{raster_obj}{object of class raster to be extracted from}

\item{summary_fun}{the function to be used in computing zonal statistics}

\item{cpus}{number of CPUs for parallelization}

\item{parallel_mode}{see \code{parallelMap::parallelLapply()} for more details}
}
\description{
A function to parallel compute zonal stats
}
\details{
This function computes the zonal statistics for a shapefile, preferably a large
shapefile with over 200,000+ observations by extracting vector object of class
\code{raster}. The function splits the \code{sf}, \code{data.frame} polygon object into \code{cpus} number
of parts and each part is extracted into in parallel.
}
